refactors and cleanups as recommended by Jason0214 https://github.com/godotengine/godot-blender-exporter/pull/252
