+++ io_scene_godot/converters/animation/serializer.py
@@ -86,9 +86,7 @@ class TransformFrame:
         # FIXME: lose negative scale
         xform_frame.scale = xform_matrix.to_scale()
 
-        if rotation_mode == 'QUATERNION':
-            xform_frame.rotation_euler = xform_matrix.to_euler()
-        elif rotation_mode == 'AXIS_ANGLE':
+        if rotation_mode in ('QUATERNION', 'AXIS_ANGLE'):
             xform_frame.rotation_euler = xform_matrix.to_euler()
         else:
             xform_frame.rotation_euler = xform_matrix.to_euler(rotation_mode)
