+++ io_scene_godot/converters/utils.py
@@ -1,7 +1,20 @@
 """Util functions and structs shared by multiple resource converters"""
+import bpy, bmesh, mathutils
 
-import bpy
-import bmesh
+gdenums = {
+    'PRIMITIVE_POINTS' : 0, # Render array as points (one vertex equals one point).
+    'PRIMITIVE_LINES'  : 1, # Render array as lines (every two vertices a line is created).
+    'PRIMITIVE_LINE_STRIP' : 2, # Render array as line strip.
+    'PRIMITIVE_LINE_LOOP' : 3, # Render array as line loop (like line strip, but closed).
+    'PRIMITIVE_TRIANGLES' : 4, # Render array as triangles (every three vertices a triangle is created).
+    'PRIMITIVE_TRIANGLE_STRIP' : 5, # Render array as triangle strips.
+    'PRIMITIVE_TRIANGLE_FAN' : 6, # Render array as triangle fans.
+}
+
+
+def fix_vertex(vtx):
+    """Changes a single position vector from y-up to z-up"""
+    return mathutils.Vector((vtx.x, vtx.z, -vtx.y))
 
 
 def get_applicable_modifiers(obj, export_settings):
